Update README.md
Adding a bit more context, especially about compatibility of v2 manifests, after I received a confused email from someone for whom it wasn't clear that the Wikimedia Commons manifest only works from OpenRefine 3.6 upwards. 

Not sure if this is the right place for documentation, but a bit of basic info won't hurt?